Add perfomance tests to the NV and WK jobs
Adding setup for the performance testing. The mocked olp server is copied from
the heremaps/here-olp-sdk-cpp github repo with the minimal changes (typescript SDK is not able to setup the mocked server
behavior with the proxy).

Adding two tests. The tests are sending 12 requests (getData) per second to the mocked olp server during 5min (short nv test) and 10hours (wk test)
 and counts successes and failed requests and memory usage.

After job done, the reports of memory usage, collected by heaptrack will be generated and the flame graph will be rendered to the svg
by the FlameGraph lib.

 Resolves: OLPEDGE-1178

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>