Add HttpError Class.
Adding HttpError class to be able to provide for consumers
a more usable errors from services.
Would be used in the methods to propagate error with http status code
and with a message in case if something went wrong during the request.

The HttpError class extends generic Error class from V8 and
adds property code for http status.

Adding the unit tests and setup running unit tests for the dataservice-api component.

Resolves: OLPEDGE-1433
Signed-off-by: ashevchu <ext-andrii.shevchuk@here.com>