Make "datalevel" is not required in statistics API
Datalevel is not required now, de to changes in the statistics service.
When using API to retrieve assets, if no data level is provided,
you get back assets at the deepest data level this layer
is configured for.

Also, change the type of datalevel from string to number (supports levels 0-11)
and add one more test to check methods if no datalevel provided.

Resolves OLPSUP-9319

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>