Tests for versioned layer client
Adding a few unit tests for testing versioned layer client
with version 0 in the constructor.

The first test checks if the version 0 have been set to the instance.
The second test checks if the method getPartitions calls service for partitions of the 0 version of the layer and shows how we can mock some class used in the testing method.
3rd test checks if the getData method does not override the version of the instance if version is present and also shows how we can mock some private method.

Relates-To: OLPEDGE-1617

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>