Fix for version is incorrectly handled issue.
This CR fixes issues with incorrectly handled version 0.
The issue was about type checking and 0 was ignoring.
The fix changes checking from if (!version) to if (version === undefined)
and this fix alows to use version 0 for the requests.

CR also adds support for setting version into the request classes.
If the user sets the version to the request, then the locked version of the instance is ignoring.
If the user does not set the version to the constructor or to the request instance, the latest version
will be fetched, used and locked to the instance of the client.

Adding integration tests to handle all checking:

- Should use version 0 for getPartitions with
PartitionsRequest.withVersion(0).withPartitionIds(['23605706']);.
- Should use version 0 for getPartitions with
QuadKeyPartitionsRequest.withVersion(0).withQuadKey(quadKeyFromMortonCode(23605706));
- Should use version 0 for getData with
DataRequest.withVersion(0).withQuadKey(quadKeyFromMortonCode(23605706));
- Should use version 0 for getData with
DataRequest.withVersion(0).withPartitionId('23605706');
- Should use locked version 142 for getPartitions with
PartitionsRequest.withPartitionIds(['23605706']);
- Should use locked version 142 for getPartitions with
QuadKeyPartitionsRequest.withQuadKey(quadKeyFromMortonCode(23605706));
- Should use locked version 142 for getData with
DataRequest.withQuadKey(quadKeyFromMortonCode(23605706));
- Should use locked version 142 for getData with
DataRequest.withPartitionId('23605706');
- Should use latest version for getPartitions with
PartitionsRequest.withPartitionIds(['23605706']) and lock it;
- Should use latest version for getPartitions with
QuadKeyPartitionsRequest.withQuadKey(quadKeyFromMortonCode(23605706)) and lock it;
- Should use latest version for getData with
DataRequest.withQuadKey(quadKeyFromMortonCode(23605706)) and lock it;
- Should use latest version for getData with
DataRequest.withPartitionId('23605706') and lock it;
- Should use version 0 for getLayerVersions with
LayerVersionsRequest.withVersion(0);
- Should use version 0 for getVersions with
CatalogVersionRequest.withEndVersion(0);

Resolves: OLPSUP-9571

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>