Add integration tests for testing API breaks for VersionedLayerClient.
Add skeleton for testing api breaks.
We cover with test the public APIs of our SDK to make sure that in the future that changes in the code will not create breaking changes and will not fail the build processes in customers side. The tests do not verify anything of the functional part, except whether our code is complied with, using the possible variants of use of the public APIs.

Add integration tests for testing API breaks for VersionedLayerClient:
* VersionedLayerClientParams with all required params
* VersionedLayerClientParams with all required and optional params
* LayerClient shoud be initialized with arguments
* LayerClient shoud be initialized with VersionedLayerClientParams
* Test getPartitions method with QuadKeyPartitionsRequest
* Test getPartitions method with QuadKeyPartitionsRequest and abort signal
* Test getPartitions method with PartitionsRequest
* Test getPartitions method with PartitionsRequest and abort signal
* Test getData method method with dataHandle
* Test getData method method with dataHandle and abort signal

Relates-To: OLPEDGE-1690

Signed-off-by: Drapak Iryna Angelica <ext-iryna.drapak@here.com>