Caching partitions metadata in memory.
This CR adds LRU cache support for the partitions metadata for the
versioned and volatile layers.
The user is able to set for each partition request param:
"online-only" and "online if not found in cache".

By default for each request uses mode "online if not found in cache".
This mode decreases the count of requests to the services.

Disabling some tests. The tests will be enabled on the next CRs.

Resolves: OLPEDGE-1667

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>