getTile() with quad tree caching.
This CR fixes the issue with the fetching quad tree index
with depth 0.

Fetching quad tree index for parent Tile of the requested Tile
with delta 4 is useful for caching all sub quads
and parent quads metadata for the later calls.

Returns the blob of the requested Tile or the blob
of the closest parent Tile.

Resolves: OLPEDGE-1953

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>