Add CheckDataExists method.
When submitting a blob for publication the user should have
the possibility to provide it's own data handle.

Data handles must be unique within the layer across all
versions and should be checked against Blob API before
used if it does not exists yet.
In case data handle exists, a new one needs to be generated
and checked again until one is found that is not present
in the blob store.

To give user the possibility to generate his own data handle
and to check that it is not used we exposing a method
`CheckDataExists` anf request class `CheckDataExistsRequest`.

Also fixing crash in the function `checkBlobExists` of blob-api
and adding unit tests.

Relates-To: OLPEDGE-2047

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>