Add a retry logic in case datahandle exist.
If datahandle was not provided, we generate the UUID and use it
as datahandle.
A retry logic in case datahandle is already present is a loop of
3 tries and each time we generate a new UUID and retry.
If all thies was not success, the method rejects the promise
with Error.

Make enum STATUS_CODES public and move to the core package to share
for all components.

Resolves: OLPEDGE-2088

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>