Add uploadBlob and uploadPartitions methods.
These two methods used for uploading data and metadata
to the service.

The uploadBlob method uploads the data in one request or
in many requests, chunk by chunk if the data is more than 50Mb.

The uploadPartitions method uploads the metadata
of the uploaded data.

Updated the method publishToBatch to reuse uploadBlob.

Moved generation of datahandle to the private method.

Added tests.

Resolves: OLPEDGE-2046, OLPEDGE-2051

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>