Add getAggregatedData to the layer clients.
Add a new `getAggregatedData` API to `VersionedLayerClient`
and `VolatileLayerClient`.
You can use it to retrieve tile data (if it exists) or the
nearest parent tile data.
Use this API for tile-tree structures where children tile data
is aggregated and stored in parent tiles.

Also deprecate params and version in `TileRequest`.

Fix issue with fetching Quad Tree Index for volatile layer.
Was fetching the latest version of catalog from Metadata service
for volatile layers.

Resolves: OLPEDGE-1893

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>