Use RequestFactory from core.
This CR mostly about changes in imports and
adaptations for tests.

As we have deprecated RequestFactory class in the
dataservice-read, we should use it from dataservice-core.

The unit tests have been updated because of starting using
RequestFactory from core makes the issues with mocking instances
in the tests.

Also the file
@here/olp-sdk-dataservice-read/lib/utils/RequestBuilderFactory.ts
was updated.

We've back the copy of class from core instead of importing from the
core, because of the circular dependency issue.
The SDK components do not use that file and that file will be removed
later as a deprecated.

Resolves: OLPEDGE-2001

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>