Caching the metadata with additional fields.
This CR fixes the issue when the user calls the first time the
metadata without additional fields and then calls the second time
the same metadata, but with additional fields.
SDK was returning from cache the metadata without additional fields.

For fixing that, we're failing cache lookup when some of additional
field was requested, but in the cache the cached partition has not that
field.

Resolves: OLPSUP-11727

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>