Fix bug with cnunks numbers.
The array of chunks for the `completeMultipartUpload`
request was incorrect and looked like this:

```
{
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 6
  },
  {
    id: '"HASH"',
    number: 12
  },
  {
    id: '"HASH"',
    number: 12
  },
  {
    id: '"HASH"',
    number: 12
  },
  {
    id: '"HASH"',
    number: 12
  },
  {
    id: '"HASH"',
    number: 12
  },
  {
    id: '"HASH"',
    number: 12
  }
```

The request was failing with error:
`400: One or more of the specified parts could not be found.
The part may not have been uploaded.`

The fix uses the chunk number from the uploaded response
instead of the counter.

Relates-To: OLPEDGE-2456

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>