Hotfix 1.11.1
This CR fixes the error on the compilation phase,
related to the type checking.
We import the module `fs` only for nodejs,
it's configured in the package.json file,
so we can disable the type checking for the importing
in the private class NodeFileData and it's safe.

Updated versions and CHANGELOG

Resolves: OLPSUP-14784

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>