Minor change in poll method.
The issue is about sending POST requests for committing ofsets
in `poll` method even if no messages in `poll` response.
The backend sends back an error about invalid parameters of
committing offsets requests in this case.

This fix checks the count of messages in the poll
response and sends committing requests only if messages are retrieved.

Relates-To: OLPEDGE-2663
Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>